REFACTOR: Moved years of experience into separate file so that it's accessible throughout the app
